feat : 챌린지 섹션 분리
