+++ igoMoney/igoMoney/Source/ChallengeScene/EmptyChallengeListSection/EmptyChallengeListSectionCore.swift
@@ -1,0 +1,21 @@
+//
+//  EmptyChallengeListSectionCore.swift
+//  igoMoney
+//
+//  Copyright (c) 2023 Minii All rights reserved.
+
+import ComposableArchitecture
+
+struct EmptyChallengeListSectionCore: Reducer {
+    struct State: Equatable {
+        
+    }
+    
+    enum Action: Equatable {
+        
+    }
+    
+    func reduce(into state: inout State, action: Action) -> Effect<Action> {
+        return .none
+    }
+}
